avi_cms_backend/requirements.txt
asgiref==3.4.1
blessings==1.7
certifi==2021.10.8
charset-normalizer==2.0.9
curtsies==0.3.10
cwcwidth==0.1.5
Django==4.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
psycopg2-binary==2.9.2
Pygments==2.10.0
python-dotenv==0.19.2
pyxdg==0.27
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.1
urllib3==1.26.7
